+++ Shared/DTOs/SubmissionDTOs/SubmissionHistoryDTO.cs
@@ -6,8 +6,8 @@ namespace Shared.DTOs.SubmissionDTOs
         public long Id { get; set; }
         public string Verdict { get; set; } = string.Empty;
         public string Language { get; set; } = string.Empty;
-        public int? RuntimeMs { get; set; }
-        public int? MemoryKb { get; set; }
+        public long? RuntimeMs { get; set; }
+        public long? MemoryKb { get; set; }
         public DateTime SubmittedAt { get; set; }
     }
 }
